#1901 rename variable move utils and change imports
