+++ src/backend/src/prisma/seed.ts
@@ -39,7 +39,7 @@ import { writeFileSync } from 'fs';
 import ProjectsService from '../services/projects.services';
 import { Decimal } from 'decimal.js';
 import DesignReviewsService from '../services/design-reviews.services';
-import { transformDate } from '../utils/utils';
+import { transformDate } from '../utils/datetime.utils';
 
 const prisma = new PrismaClient();
