+++ src/backend/src/utils/change-requests.utils.ts
@@ -184,12 +184,12 @@ export const validateProposedChangesFields = async (
 };
 
 /**
- * Determines if there are other open unreviewed change requests for this wbs element
+ * throws an error if there are any other open unreviewed change requests for this wbs element
  * @param wbsElemId the wbs element id to find CRs with
  * @throws if the WBS element has open unreviewed change requests
  *
  */
-export const validateUnreviewedOpenCRs = async (wbsElemId: number) => {
+export const validateNoUnreviewedOpenCRs = async (wbsElemId: number) => {
   const openCRs = await prisma.change_Request.findMany({
     where: { wbsElementId: wbsElemId, accepted: false }
   });
