+++ src/backend/src/utils/datetime.utils.ts
@@ -1,0 +1,5 @@
+export const transformDate = (date: Date) => {
+  const month = date.getMonth() + 1 < 10 ? `0${date.getMonth() + 1}` : (date.getMonth() + 1).toString();
+  const day = date.getDate() < 10 ? `0${date.getDate()}` : date.getDate().toString();
+  return `${date.getFullYear().toString()}-${month}-${day}`;
+};
