+++ src/backend/src/utils/utils.ts
@@ -23,9 +23,3 @@ export const validateInputs = (req: Request, res: Response, next: Function): Res
   }
   next();
 };
-
-export const transformDate = (date: Date) => {
-  const month = date.getMonth() + 1 < 10 ? `0${date.getMonth() + 1}` : (date.getMonth() + 1).toString();
-  const day = date.getDate() < 10 ? `0${date.getDate()}` : date.getDate().toString();
-  return `${date.getFullYear().toString()}-${month}-${day}`;
-};
