+++ src/frontend/src/utils/datetime.utils.ts
@@ -17,3 +17,9 @@ export const dateToString = (date: Date) => {
 export const dateFormatMonthDate = (date: Date) => {
   return dayjs(date).format('MMM D');
 };
+
+export const transformDate = (date: Date) => {
+  const month = date.getMonth() + 1 < 10 ? `0${date.getMonth() + 1}` : (date.getMonth() + 1).toString();
+  const day = date.getDate() < 10 ? `0${date.getDate()}` : date.getDate().toString();
+  return `${date.getFullYear().toString()}/${month}/${day}`;
+};
