+++ src/frontend/src/utils/task.utils.ts
@@ -56,12 +56,6 @@ export interface TaskListDataGridProps {
   setDisabled: Dispatch<SetStateAction<boolean>>;
 }
 
-export const transformDate = (date: Date) => {
-  const month = date.getMonth() + 1 < 10 ? `0${date.getMonth() + 1}` : (date.getMonth() + 1).toString();
-  const day = date.getDate() < 10 ? `0${date.getDate()}` : date.getDate().toString();
-  return `${date.getFullYear().toString()}/${month}/${day}`;
-};
-
 export const taskUserToAutocompleteOption = (user: User): { label: string; id: number } => {
   return { label: `${fullNamePipe(user)} (${user.email})`, id: user.userId };
 };
